Move ssrf to public vulnerabilities package
