+++ sample-apps/echo-v5-postgres/go.mod
@@ -14,7 +14,8 @@ require (
 	github.com/jtolds/gls v4.20.0+incompatible // indirect
 	github.com/tetratelabs/wazero v1.11.0 // indirect
 	go4.org/netipx v0.0.0-20231129151722-fdeea329fbba // indirect
-	golang.org/x/sys v0.40.0 // indirect
+	golang.org/x/sys v0.41.0 // indirect
+	golang.org/x/text v0.34.0 // indirect
 )
 
 replace github.com/AikidoSec/firewall-go => ../../
