+++ vulnerabilities/ssrf/vulnerability.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 
 	"github.com/AikidoSec/firewall-go/internal/agent/ipaddr"
-	"github.com/AikidoSec/firewall-go/internal/vulnerabilities"
+	"github.com/AikidoSec/firewall-go/vulnerabilities"
 )
 
 // SSRFVulnerability detects server-side request forgery.
